[feat] url maintenance (#53)
This PR maintains the user's path URL when checking for logged-in status.

Previously, whenever user first load a page with a path (e.g. /mates), the app redirects them to login page and then to the dashboard after verifying their auth.
If they are logged out, it still leads user back to dashboard.