[feat] Create mate backend functionalities (#49)
## Changes made

This PR adds functionalities to the backend so that mates can be added. The frontend communicates this function via GraphQL, and that has also been added. Only the types that are returnable has been added to the GQL schema for now (i.e. user is not typed yet). 

Chain PRs:
1. **backend [THIS]** #49 
2. frontend #50 
3. migration script #51

New Models:
- Mate
- Family
- Sync

For backward compatibility purposes, a new `unassigned_family_id` is created and added if it does not exist in the current user.


## Screencapture (if applicable)

https://user-images.githubusercontent.com/12789302/138609928-7cf33369-2b42-49a7-8f52-ea3fbd2791ed.mov



## Work left to be done


Co-authored-by: Piyush Kancharlawar <piyushgk1@gmail.com>