[fix] Absolute graphql location to fix dev/ci pytest differences (#115)
## Changes made
In #108 the dockerfile was changed to use the `/home/worker/app` work directory instead of the previous '/home/worker/app/server` directory only during development. 

This caused the graphql import in `api.py` to fail since it was a relative import that assumed the app would be started from within the `server/` folder. 

The easy/simple fix is to replace the relative import with an absolute import. This, however fails in ci since ci does not use the docker images and rather runs directly on gh action runners.

Thus, in order to enact this solution, we have to migrate the backend ci action to use the devbox container.

Additionally we use the `bin/dev` commands instead of direct shell commands in order to reduce code duplication and drift between the dev environment and the ci environment.

## Screencapture (if applicable)


## Testing
- [ ] End-to-End


## Work left to be done