[feat] Move pre-commit action to use bin/dev commands (#126)
## Changes made
Change the pre-commit github action to use devbox container instead of running the commands directly on gh action runner. 

This is an extension to #115 that helps standardize our github actions and allows python upgrade in #69 to proceed more smoothly.

Pre-commit installs all dependencies by itself (pip and npm dependencies for each pre-commit hook). It however, runs into a npm root user error.
![image](https://user-images.githubusercontent.com/9906905/147390416-4beb05bc-fc9b-4ce2-8290-4aa101de82a8.png)

As a workaround, we run pre-commit as root. 

This is the same issue that @leedm98 ran into when setting up pre-commit, and I was able to consistently repro it by deleting the `.pre-commit` folder and trying to install from a clean workspace. Added the workaround as a setup step in the readme.

## Screencapture (if applicable)

## Testing
- [ ] End-to-End


## Work left to be done
Figure out why pre-commit has npm root user issues, and fix configuration to avoid the use of super user permissions.